code/requirements.txt
tqdm==4.8.4
Pillow==4.2.0 # nice
